Check database status before CRUD.

+++ app/src/main/kotlin/com/absinthe/libchecker/database/LCRepository.kt
@@ -8,6 +8,7 @@ import com.absinthe.libchecker.database.entity.SnapshotItem
 import com.absinthe.libchecker.database.entity.TimeStampItem
 import com.absinthe.libchecker.database.entity.TrackItem
 import kotlinx.coroutines.flow.Flow
+import timber.log.Timber
 
 class LCRepository(private val lcDao: LCDao) {
 
@@ -15,7 +16,16 @@ class LCRepository(private val lcDao: LCDao) {
   val allSnapshotItemsFlow: Flow<List<SnapshotItem>> =
     lcDao.getSnapshotsFlow(GlobalValues.snapshotTimestamp)
 
+  private fun checkDatabaseStatus(): Boolean {
+    if (LCDatabase.isClosed()) {
+      Timber.w("Database is closed")
+      return false
+    }
+    return true
+  }
+
   suspend fun getItem(packageName: String): LCItem? {
+    if (checkDatabaseStatus().not()) return null
     return lcDao.getItem(packageName)
   }
 
@@ -28,54 +38,67 @@ class LCRepository(private val lcDao: LCDao) {
   suspend fun getTrackItems(): List<TrackItem> = lcDao.getTrackItems()
 
   suspend fun insert(item: LCItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insert(item)
   }
 
   suspend fun insert(list: List<LCItem>) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insert(list)
   }
 
   suspend fun insert(item: SnapshotItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insert(item)
   }
 
   suspend fun insert(item: TrackItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insert(item)
   }
 
   suspend fun insert(item: TimeStampItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insert(item)
   }
 
   suspend fun insertSnapshots(items: List<SnapshotItem>) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insertSnapshots(items)
   }
 
   suspend fun update(item: LCItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.update(item)
   }
 
   suspend fun update(item: SnapshotItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.update(item)
   }
 
   suspend fun update(items: List<SnapshotItem>) {
+    if (checkDatabaseStatus().not()) return
     lcDao.update(items)
   }
 
   suspend fun delete(item: SnapshotItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.delete(item)
   }
 
   suspend fun delete(item: LCItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.delete(item)
   }
 
   suspend fun delete(item: TrackItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.delete(item)
   }
 
   suspend fun deleteSnapshotsAndTimeStamp(timestamp: Long) {
+    if (checkDatabaseStatus().not()) return
     val list = getSnapshots(timestamp)
     var count = 0
     val chunk = mutableListOf<SnapshotItem>()
@@ -97,30 +120,37 @@ class LCRepository(private val lcDao: LCDao) {
   }
 
   suspend fun updateTimeStampItem(item: TimeStampItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.update(item)
   }
 
   fun deleteAllSnapshots() {
+    if (checkDatabaseStatus().not()) return
     lcDao.deleteAllSnapshots()
   }
 
   fun deleteAllItems() {
+    if (checkDatabaseStatus().not()) return
     lcDao.deleteAllItems()
   }
 
   suspend fun insertSnapshotDiffItems(item: SnapshotDiffStoringItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.insertSnapshotDiff(item)
   }
 
   suspend fun updateSnapshotDiff(item: SnapshotDiffStoringItem) {
+    if (checkDatabaseStatus().not()) return
     lcDao.updateSnapshotDiff(item)
   }
 
   suspend fun deleteSnapshotDiff(packageName: String) {
+    if (checkDatabaseStatus().not()) return
     lcDao.deleteSnapshotDiff(packageName)
   }
 
   fun deleteAllSnapshotDiffItems() {
+    if (checkDatabaseStatus().not()) return
     lcDao.deleteAllSnapshotDiffItems()
   }
 
@@ -128,10 +158,12 @@ class LCRepository(private val lcDao: LCDao) {
     lcDao.getSnapshotDiff(packageName)
 
   fun updateKotlinUsage(packageName: String, used: Boolean) {
+    if (checkDatabaseStatus().not()) return
     lcDao.updateKotlinUsage(packageName, used)
   }
 
   fun updateKotlinUsage(map: Map<String, Boolean>) {
+    if (checkDatabaseStatus().not()) return
     lcDao.updateKotlinUsage(map)
   }
 }
